Change LinkSpec position/length fields from int to qsizetype to match Qt string and regex APIs

+++ src/Console/DummyParser.cpp
@@ -39,7 +39,7 @@ LineParser::Result DummyParser::handleLine(const QString &message,
     file = file.trimmed();
     const QFileInfo fileInfo{file};
     LinkSpec link{
-        static_cast<int>(regExpMatch.capturedStart()), static_cast<int>(regExpMatch.capturedLength()),
+        regExpMatch.capturedStart(), regExpMatch.capturedLength(),
         addLinkSpecForAbsoluteFilePath(fileInfo.absoluteFilePath(), "-1")};
     return Result{Status::Done, message, {link}};
   }
